feat: Add onembed_image method to GyazoClient
The `onembed_image` method is added to the `GyazoClient` struct to embed an image using the Gyazo API. This method takes a URL as input and returns an `OembedResponse` struct. The URL must start with 'https://gyazo.com/' for the method to work properly. If the URL is invalid, an `InvalidUrl` error is returned.